Some more tweeks to get the Engine and GFX interfacing correctly.
Fixed some issues with physics.  There still seems to be some trouble with DNA execution in the program.

Reverted some of Sprotiel's changes:
1.  Inlines.  Inlining is important for speed in many cases (I've checked, I promise).  If they're causing errors and not just warnings we can figure out something.  Otherweise maybe just changing what warnings are displayed in Dev-C++ is the way to go.

2.  ios::nocreate is only defind in iostream.h (not iostream).  It's important this is there so the program can detect if no such file exists.  Without this the program doesn't operate the way in which it should.

3.  The order of #include <gl/gl.h> etc. is important.  There are things defined in GUIMain.h which must be included before you include the OpenGL libraries.

Any other reverts may just be an oversight on my part.

+++ Engine/DNA_Execution.cpp
@@ -53,7 +53,7 @@ void ExecuteLogic(int n);
 
 __int32 StackCeil(float value)
 {
-    if (value < -2147483648.)
+    if (value < -2147483648)
         return -2147483648;
     else
         return 2147483647;
@@ -439,8 +439,8 @@ void finddist()
 {
     float a,b,c;
 
-    b = currbot->y() - PopIntStack();
-    a = currbot->x() - PopIntStack();
+    b = float(long(currbot->y()) - PopIntStack());
+    a = float(long(currbot->x()) - PopIntStack());
     c = sqrt(b*b+a*a);
 
     PushIntStack(StackCeil(c));
